modified

+++ src/STRIVER_SHEET/PATTERNS/Pattern13.java
@@ -7,9 +7,9 @@ public class Pattern13 {
 
     public static void nNumberTriangle(int n) {
         int current=1;
-        for(int i=1;i<=n;++i)
+        for(byte  i=1;i<=n;++i)
         {
-            for(int j=1;j<=i;++j)
+            for(byte j=1;j<=i;++j)
             {
                 System.out.print(current + " ");
                 ++current;
